pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "muenster-bus-board"
version = "0.1.0"
description = "Desktop live board for Muenster bus departures and vehicle positions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pywebview>=5.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
